Resolve "Release 4.2.1"

See merge request maja/maja!147

+++ orchestrator/cots/otb/otb_file_utils.py
@@ -1,4 +1,19 @@
 # -*- coding: utf-8 -*-
+#
+# Copyright (C) 2020 Centre National d'Etudes Spatiales (CNES)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 """
 ###################################################################################################
 
@@ -16,11 +31,6 @@ orchestrator.common.file_utils -- shortdesc
 orchestrator.common.file_utils is a description
 
 It defines classes_and_methods
-
-###################################################################################################
-
-:copyright: 2019 CNES. All rights reserved.
-:license: license
 
 ###################################################################################################
 """
@@ -49,14 +59,15 @@ def otb_copy_file_to_directory(sourceFilename, destinationDir):
     return destinationFilename
 
 
-def otb_copy_file(source, dest, raise_exceptions=True):
+def otb_copy_image_to_file(source, dest, raise_exceptions=True):
     if not get_test_mode():
         if type(source).__name__ == "SwigPyObject":
             write_images([source], [dest])
         else:
             LOGGER.debug("source : " + source + " , dest : " + dest)
             try:
-                shutil.copyfile(source, dest)
+                if not os.path.exists(dest) or not os.path.samefile(source, dest):
+                    shutil.copyfile(source, dest)
             except IOError as err:
                 if raise_exceptions:
                     raise MajaIOError(err)
@@ -66,3 +77,7 @@ def otb_copy_file(source, dest, raise_exceptions=True):
     else:
         with open(dest, 'a'):
             os.utime(dest, None)
+
+
+def otb_is_swig_pointer(im):
+    return type(im).__name__ == "SwigPyObject"
